Corrected Issue with Opening Recent Entries
Line 332 in main.py issues the command to 'open_xml_file', but does this with event.GetId().
The getId() method gets the id of the event - not the id of the object that created the event.

Hence changed to GetInt(), which gets the id that called the event (in this case the menuitem on recent entries).