Working! Although inefficient
Took away 'add node' button from all except the main node. This makes it work, but reduces capability.

Added ability to change the xml tag, as well as the text value. Didn't do this well.
I need to figure out a better way to link 'view' objects to data, so that when they are changed, the view updates.
Currently the xml tree view wont change, even if you change a tag.